Simplificar lógica y añadir controladores nuevos
Se simplificó la lógica de retorno en `BuscarClienteController.cs`.
Se cambiaron tipos de datos de `int` a `decimal` en las propiedades
`NCantidadCompra` y `NCantidadVenta` en los archivos `Compra.cs`,
`Venta.cs` y su configuración en `DbNegocioContext.cs`.

Se añadieron las propiedades `IdPersona` y `CDocumento` al modelo
`UsuarioLogin` y se actualizó su uso en `SolicitudCreditoService.cs`.

Se implementaron dos nuevos controladores:
- `AuthController` para autenticación de usuarios con manejo de
  intentos fallidos y bloqueos.
- `CambioProductoController` para consultar opciones de cambio
  de producto.

Finalmente, se añadió el modelo `LoginRequest` para estructurar
las solicitudes de inicio de sesión.

+++ CrediAvanzaAPI/Models/Compra.cs
@@ -9,7 +9,7 @@ public partial class Compra
 
     public string CProducto { get; set; } = null!;
 
-    public int NCantidadCompra { get; set; }
+    public decimal NCantidadCompra { get; set; }
 
     public int NUnidadMedida { get; set; }
